refactoring & bug fix
